chore: move some utils from update tests for reuse
